Move 'Delete Account' button from ProfileActivity to EditProfileActivity
