+++ app/src/main/res/layout-sw320dp/activity_edit_profile.xml
@@ -97,6 +97,18 @@
         app:layout_constraintTop_toBottomOf="@+id/edit_text_edit_profile_phone"
         android:autofillHints="" />
 
+    <Button
+        android:id="@+id/button_delete_account"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="20dp"
+        android:onClick="deleteUserAccount"
+        android:text="@string/activity_profile_icon_delete_account"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/edit_text_edit_profile_description" />
+
+
     <ProgressBar
         android:id="@+id/progress_bar_edit_profile"
         android:layout_width="wrap_content"
